pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmhub"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn",
  "httpx",
  "openai",
  "pydantic",
  "python-dotenv",
  "google-genai",
  "typer",
]

[project.scripts]
llmhub = "cli:app"

[tool.setuptools]
packages = ["app"]
py-modules = ["cli"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
